// src/yalps/types.ts
export type Constraint = {
  /**
   * The total should be equal to this number.
   * In the case that `min` or `max` are also defined, this is used instead.
   */
  readonly equal?: number

  /**
   * The total should be greater than or equal to this number.
   * Can be specified alongside `max`.
   */
  readonly min?: number

  /**
   * The total should be less than or equal to this number.
   * Can be specified alongside `min`.
   */
  readonly max?: number
}

/**
 * The coefficients of a variable represented as either an object or an `Iterable`.
 * `ConstraintKey` should extend `string` if this is an object.
 * If this is an `Iterable` and has duplicate keys, then the last entry is used for each set of duplicate keys.
 */
export type Coefficients<ConstraintKey = string> =
  | Iterable<readonly [ConstraintKey, number]>
  | ([ConstraintKey] extends [string] ? Readonly<Partial<Record<ConstraintKey, number>>> : never)

/**
 * Indicates whether to maximize or minimize the objective.
 */
export type OptimizationDirection = "maximize" | "minimize"
// ...
export type Model<VariableKey = string, ConstraintKey = string> = {
  /**
   * Indicates whether to `"maximize"` or `"minimize"` the objective.
   * Defaults to `"maximize"` if left blank.
   */
  readonly direction?: OptimizationDirection

  /**
   * The key of the value to optimize. Can be omitted,
   * in which case the solver gives some solution (if any) that satisfies the constraints.
   * @example
   * Note that constraints can be placed upon the objective itself.
   * Maximize up to a certain point:
   * ```
   * const model = {
   *   direction: "maximize",
   *   objective: "obj",
   *   constraints: { obj: { max: 100 } },
   *   variables: [
   *     // ...
   *   ]
   * }
   * ```
   */
  readonly objective?: ConstraintKey

  /**
   * An object or `Iterable` representing the constraints of the problem.
   * In the case `constraints` is an `Iterable`, duplicate keys are not ignored.
   * Rather, the bounds on the constraints are merged to become the most restrictive.
   * @see {@link Constraint}
   * @example
   * Constraints as an object:
   * ```
   * const constraints = {
   *   a: { max: 7 },
   *   b: { equal: 22 },
   *   c: { min: 5 }
   * }
   * ```
   * @example
   * Constraints as an `Iterable`:
   * ```
   * const constraints =
   *   new Map<string, Constraint>()
   *     .set("a", { max: 7 })
   *     .set("b", { equal: 22 })
   *     .set("c", { min: 5 })
   * ```
   */
  readonly constraints:
    | Iterable<readonly [ConstraintKey, Constraint]>
    | ([ConstraintKey] extends [string] ? Readonly<Partial<Record<ConstraintKey, Constraint>>> : never)

  /**
   * An object or `Iterable` representing the variables of the problem.
   * In the case `variables` is an `Iterable`, duplicate keys are not ignored.
   * The order of variables is preserved in the solution,
   * but variables that end up having a value of `0` are not included in the solution by default.
   * @example
   * Variables as an object:
   * ```
   * const variables = {
   *   x: { a: 2, b: 11 },
   *   y: { a: 3, c: 22 }
   * }
   * ```
   * @example
   * Variables as an `Iterable`:
   * ```
   * const variables =
   *   new Map<string, Coefficients>()
   *     .set("x", { a: 2, b: 11 })
   *     .set("y", { a: 3, c: 22 })
   * ```
   * @example
   * Mixing objects and `Iterable`s:
   * ```
   * const variables = {
   *   x: { a: 2, b: 11 },
   *   y: [["a", 3], ["c", 22]]
   * }
   * ```
   */
  readonly variables:
    | Iterable<readonly [VariableKey, Coefficients<ConstraintKey>]>
    | ([VariableKey] extends [string] ? Readonly<Partial<Record<VariableKey, Coefficients<ConstraintKey>>>> : never)

  /**
   * An `Iterable` of variable keys that indicate the corresponding variables are integer.
   * It is recommended to use a {@link Set} as the `Iterable`.
   * It can also be a `boolean`, indicating whether all variables are integer or not.
   * If this is left blank, then all variables are treated as not integer.
   */
  readonly integers?: boolean | Iterable<VariableKey>

  /**
   * An `Iterable` of variable keys that indicate the corresponding variables are binary
   * (can only be 0 or 1 in the solution).
   * It is recommended to use a {@link Set} as the `Iterable`.
   * It can also be a `boolean`, indicating whether all variables are binary or not.
   * If this is left blank, then all variables are treated as not binary.
   */
  readonly binaries?: boolean | Iterable<VariableKey>
}